rpi/requirements.txt
# Drone Client Requirements
# Core dependencies
fastapi==0.119.0
uvicorn==0.37.0
websockets==15.0.1
pymavlink==2.4.37
opencv-python==4.8.1.78
numpy==1.24.3
pyyaml==6.0.1
smbus2==0.4.3
RPi.GPIO==0.7.1
picamera2==0.3.12
ultralytics==8.0.196
onnxruntime==1.16.3

# Optional VLM dependencies
aiohttp==3.9.1
google-generativeai==0.3.2

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
